requirements.txt
flask
flask-cors
requests
pillow
gunicorn